requirements.txt
# --- Kern-Funktionalität ---
requests==2.32.5
PyYAML==6.0.3

# --- TUI & Interface (Das "Gesicht" von 2.0) ---
textual==8.0.0
rich==14.3.3
questionary==2.1.1

# --- Metriken & Utilities ---
tiktoken==0.12.0
python-dotenv==1.2.1